# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Docs MlCity Project",
        "description": "API MlCity",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/sights": {
            "get": {
                "tags": [
                    "Sight"
                ],
                "summary": "Get all sights by filters",
                "operationId": "7a0f2b46a70e5704c7fd004ac47df360",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Sight name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "statuses",
                        "in": "query",
                        "description": "Statuses sight",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "statusLast",
                        "in": "query",
                        "description": "True = last statuses event. False = search by all statuses sight",
                        "schema": {
                            "type": "bool"
                        }
                    },
                    {
                        "name": "sightTypes",
                        "in": "query",
                        "description": "Type sight",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "description": "Address sight",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sponsor",
                        "in": "query",
                        "description": "Sponsor sight",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "searchText",
                        "in": "query",
                        "description": "Search text by name event, sponsor event and description event",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user",
                        "in": "query",
                        "description": "Email and name user author",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "region",
                        "in": "query",
                        "description": "Region sight",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "City sight",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "latitude",
                        "in": "query",
                        "description": "Latitude sight",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "description": "Longitude sight",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "not authentication"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/sights/update-vk-likes": {
            "post": {
                "tags": [
                    "Sight"
                ],
                "summary": "Update vk likes sight",
                "operationId": "29d5b359f2596b6bcaf60c73befd4a45",
                "parameters": [
                    {
                        "name": "sight_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "likes_count",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "not authentication"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/sights/set-sight-user-liked": {
            "post": {
                "tags": [
                    "Sight"
                ],
                "summary": "Set sight user liked",
                "operationId": "150c1787679a9330d64fb8ad5ac98596",
                "parameters": [
                    {
                        "name": "sight_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "not authentication"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/sights/{id}/check-user-liked": {
            "post": {
                "tags": [
                    "Sight"
                ],
                "summary": "Check user liked",
                "operationId": "037fa3e5577a365ba1893b6eebffb7b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "not authentication"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/sights/{id}/check-user-favorite": {
            "post": {
                "tags": [
                    "Sight"
                ],
                "summary": "Check user favorite",
                "operationId": "f4fefa3e9aec439c2b8afe15a12635be",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "not authentication"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/sights/{id}": {
            "get": {
                "tags": [
                    "Sight"
                ],
                "summary": "Get sight by id",
                "operationId": "7947b0b8c3c2597f8ffdf027b9d64824",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "not authentication"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}